+++ app/__init__.py
@@ -2,5 +2,5 @@
 
 from __future__ import annotations
 
-APP_VERSION = "180926.2"
+APP_VERSION = "180926.3"
 __version__ = APP_VERSION